frontend/jupter: refactor get_kernels_by_name_or_language
